refactor: added reexport to components
