--- src/components/index.ts
@@ -1,8 +1,0 @@
-import { Button } from './button/button';
-import { Container } from './container/container';
-import { Input } from './input/input';
-import { Layout } from './layout/layout';
-import { Link } from './link/link';
-import { Tooltip } from './tooltip/tooltip';
-
-export { Button, Container, Input, Layout, Link, Tooltip };
