+++ src/components/input/index.ts
@@ -1,0 +1,1 @@
+export { Input } from './input';
