+++ src/modules/main-page-content/components/main-benefits/components/benefit-item/benefit-item.tsx
@@ -1,4 +1,4 @@
-import { Tooltip } from '@/components/tooltip/tooltip';
+import { Tooltip } from '@/components/tooltip';
 import type { BenefitItemType } from '@/types/benefits';
 
 import styles from './benefit-item.module.scss';
