+++ src/modules/main-page-content/components/main-benefits/components/benefit-item/index.ts
@@ -1,0 +1,1 @@
+export { BenefitItem } from './benefit-item';
