+++ src/modules/main-page-content/components/main-form/components/contact-form/index.ts
@@ -1,0 +1,1 @@
+export { ContactForm } from './contact-form';
