+++ src/modules/main-page-content/components/main-promo/main-promo.tsx
@@ -1,5 +1,5 @@
 import promoGirlImg from '@/assets/images/student.png';
-import { Button } from '@/components/button/button';
+import { Button } from '@/components/button';
 import { Container } from '@/components/container';
 
 import styles from './main-promo.module.scss';
